Added keyboard support.
Replaced the ms-choice div with a button that opens the multiple select
drop down. Select options can be navigated with the tab key and the drop
down can be closed with either the esc key or tabbing past all available
select options. Select optoins can be selected with the space key. Also
added a checkbox for option group labels so they can be selected using
the keyboard.