Added 3 new options: 'allSelected', 'minumimCountSelected' and 'countSelected'
'allSelected' can be used to overwrite the text shown when all the items
are selected.
'minumimCountSelected' and 'countSelected' can be used to show a
formatted text after X items where selected. For example: '3 of 10
selected' will be shown if 'minumimCountSelected' is set to 3 and
'countSelected' is set to '# of % selected'.