Adding blockSeperator option
I needed for a project something similar to the optgroup function - but without the checkbox, just a plain li element to seperate several groups of options. So I added a "blockSeperator" option. Setting this to a string value causes all options elements set to this value to be a plain li element with the CSS class set to the blockSeperator value.
Easily saying: <option value="justAHeadline">Text</option> results (with a blockSeperator set to "justAHeadline") an element <li><label class="JustAHeadline">Text</label></li> instead to a checkbox block.
Maybe you think this is useful ... just wanted to contribute it :)